# Speed up smooth draw.