From 

```
   min time    avg time  std dv   runs name
   6.986 ms    7.058 ms  ±0.117  x2195 smooth x-translate
   6.982 ms    7.072 ms  ±0.215  x1435 smooth y-translate
   7.965 ms    8.871 ms  ±1.819  x3148 smooth translate
   8.260 ms   10.866 ms  ±3.096  x1398 smooth rotate 45
```

to 

```
   4.274 ms    4.367 ms  ±0.144  x1274 smooth x-translate
   4.621 ms    4.660 ms  ±0.022  x2372 smooth y-translate
   6.966 ms    7.015 ms  ±0.042  x1938 smooth translate
   7.096 ms    7.518 ms  ±0.631  x1250 smooth rotate 45
```